Change window orientation change condition
Make the window and screen(wayland output) orientation in sync unless
WEBOS_WINDOW_NO_AUTO_ORIENTATION is set.
If the condition is as below (#0 and #1|#2), change window's geometry.
0) WEBOS_WINDOW_NO_AUTO_ORIENTATION is not set as 1.
1) window is portrait and screen is landscape.
2) window is landscape and screen is portrait.
We don't rely on the window state any more for the orientation change.
setWindowStateInternal is called in WebOSPlatformWindow constructor as
we removed it from the ancestor's.

Change-Id: I1dceba187c15ee72c6a41b4f59939a4bfff4e038